Utveklser idporten token med tokenx.
Utveksler token fra idporten med tokenx, scopet for k9-mellomlagring.
