+++ nais/dev-gcp.json
@@ -24,7 +24,8 @@
     "K9_SELVBETJENING_OPPSLAG_TOKENX_AUDIENCE": "dev-fss:dusseldorf:k9-selvbetjening-oppslag",
     "K9_MELLOMLAGRING_SERVICE_DISCOVERY": "http://k9-mellomlagring",
     "K9_MELLOMLAGRING_INGRESS": "https://k9-mellomlagring.dev.intern.nav.no/v1/dokument",
-    "K9_MELLOMLAGRING_CLIENT_ID": "8d0460bd-ea81-4b08-b524-cf04874a794c/.default"
+    "K9_MELLOMLAGRING_CLIENT_ID": "8d0460bd-ea81-4b08-b524-cf04874a794c/.default",
+    "K9_MELLOMLAGRING_TOKENX_AUDIENCE": "dev-gcp:dusseldorf:k9-mellomlagring"
   },
   "slack-channel": "sif-alerts-dev",
   "slack-notify-type": "<!here> | k9-ettersending-api | ",
