+++ src/main/kotlin/no/nav/k9/Configuration.kt
@@ -41,7 +41,9 @@ data class Configuration(val config : ApplicationConfig) {
 
     internal fun getK9MellomlagringUrl() = URI(config.getRequiredString("nav.gateways.k9_mellomlagring_url", secret = false))
     internal fun getK9MellomlagringScopes() = getScopesFor("k9-mellomlagring-client-id")
+    fun getK9MellomlagringTokenxAudience(): Set<String> = getScopesFor("k9_mellomlagring_tokenx_audience")
     internal fun getK9MellomlagringIngress() = URI(config.getRequiredString("nav.gateways.k9_mellomlagring_ingress", secret = false))
+
     fun getK9SelvbetjeningOppslagTokenxAudience(): Set<String> = getScopesFor("k9_selvbetjening_oppslag_tokenx_audience")
 
     private fun getScopesFor(operation: String) = config.getRequiredList("nav.auth.scopes.$operation", secret = false, builder = { it }).toSet()
