+++ src/main/resources/application.conf
@@ -62,6 +62,7 @@ nav {
     scopes = {
       k9-mellomlagring-client-id = ""
       k9-mellomlagring-client-id = ${?K9_MELLOMLAGRING_CLIENT_ID}
+      k9_mellomlagring_tokenx_audience = ${?K9_MELLOMLAGRING_TOKENX_AUDIENCE}
       k9_selvbetjening_oppslag_tokenx_audience = ${?K9_SELVBETJENING_OPPSLAG_TOKENX_AUDIENCE}
     }
   }
